Added new test class BMInterface000Test to force test player creation to occur first before any other BMInterface unit tests
